requirements.txt
asgiref==3.7.2
Django==4.2.3
sqlparse==0.4.4
tzdata==2023.3
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
mysqlclient==2.2.0
python-decouple==2.4
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3